channel_hash: widen substrate-wide u16 → u32

+++ go/cortex_test.go
@@ -7,7 +7,7 @@ import (
 	"time"
 )
 
-const testOrigin = uint32(0xABCDEF01)
+const testOrigin = uint64(0xABCDEF01)
 
 // ---------------------------------------------------------------------------
 // Redex + RedexFile
